// src/shared/api/generated/models/supportExternalCatalogResponseDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Retenive Backend Platform
 * CMS, integration, chat and realtime API for Retenive AI Assistant
 * OpenAPI spec version: 0.1.0
 */
import type { SupportExternalCatalogResponseDtoCatalog } from "./supportExternalCatalogResponseDtoCatalog";
import type { SupportExternalCatalogSnapshotResponseDto } from "./supportExternalCatalogSnapshotResponseDto";

export interface SupportExternalCatalogResponseDto {
  catalog?: SupportExternalCatalogResponseDtoCatalog;
  /**
   * @maxLength 1024
   * @nullable
   */
  nextCursor: string | null;
  snapshot: SupportExternalCatalogSnapshotResponseDto;
}