// src/shared/api/generated/models/supportExternalCommandStatusResponseDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Retenive Backend Platform
 * CMS, integration, chat and realtime API for Retenive AI Assistant
 * OpenAPI spec version: 0.1.0
 */
import type { SupportExternalCommandStatusResponseDtoAllowedActionsItem } from "./supportExternalCommandStatusResponseDtoAllowedActionsItem";
import type { SupportExternalCommandStatusResponseDtoErrorCategory } from "./supportExternalCommandStatusResponseDtoErrorCategory";
import type { SupportExternalCommandStatusResponseDtoIntent } from "./supportExternalCommandStatusResponseDtoIntent";
import type { SupportExternalCommandStatusResponseDtoStatus } from "./supportExternalCommandStatusResponseDtoStatus";

export interface SupportExternalCommandStatusResponseDto {
  allowedActions: SupportExternalCommandStatusResponseDtoAllowedActionsItem[];
  commandId: string;
  createdAt: string;
  /** @nullable */
  errorCategory: SupportExternalCommandStatusResponseDtoErrorCategory;
  /**
   * @maxLength 100
   * @nullable
   */
  errorCode: string | null;
  intent: SupportExternalCommandStatusResponseDtoIntent;
  /** @nullable */
  nextAttemptAt: string | null;
  /** @nullable */
  resolvedAt: string | null;
  status: SupportExternalCommandStatusResponseDtoStatus;
  /** @minimum 1 */
  version: number;
}