// src/shared/api/generated/models/supportExternalMappingPublishedRevisionResponseDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Retenive Backend Platform
 * CMS, integration, chat and realtime API for Retenive AI Assistant
 * OpenAPI spec version: 0.1.0
 */
import type { SupportExternalMappingDefinitionDto } from "./supportExternalMappingDefinitionDto";
import type { SupportExternalMappingPublishedRevisionResponseDtoPublicationKind } from "./supportExternalMappingPublishedRevisionResponseDtoPublicationKind";
import type { SupportExternalMappingPublishedRevisionResponseDtoRollbackReasonCode } from "./supportExternalMappingPublishedRevisionResponseDtoRollbackReasonCode";
import type { SupportExternalMappingPublishedRevisionResponseDtoStatus } from "./supportExternalMappingPublishedRevisionResponseDtoStatus";

export interface SupportExternalMappingPublishedRevisionResponseDto {
  /** @nullable */
  catalogSnapshotId?: string | null;
  connectionId: string;
  definition: SupportExternalMappingDefinitionDto;
  /** @maxLength 128 */
  formRevision: string;
  id: string;
  publicationKind: SupportExternalMappingPublishedRevisionResponseDtoPublicationKind;
  publishedAt: string;
  /** @minimum 1 */
  revisionNumber: number;
  /** @nullable */
  rollbackReasonCode?: SupportExternalMappingPublishedRevisionResponseDtoRollbackReasonCode;
  rootId: string;
  /** @nullable */
  sourceRevisionId?: string | null;
  status: SupportExternalMappingPublishedRevisionResponseDtoStatus;
}