// src/shared/api/generated/models/supportExternalProjectItemResponseDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Retenive Backend Platform
 * CMS, integration, chat and realtime API for Retenive AI Assistant
 * OpenAPI spec version: 0.1.0
 */
import type { SupportExternalProjectItemResponseDtoAllowedActionsItem } from "./supportExternalProjectItemResponseDtoAllowedActionsItem";
import type { SupportExternalProjectItemResponseDtoAssignee } from "./supportExternalProjectItemResponseDtoAssignee";
import type { SupportExternalProjectItemResponseDtoFreshness } from "./supportExternalProjectItemResponseDtoFreshness";
import type { SupportExternalProjectItemResponseDtoLink } from "./supportExternalProjectItemResponseDtoLink";
import type { SupportExternalProjectItemResponseDtoProvider } from "./supportExternalProjectItemResponseDtoProvider";
import type { SupportExternalProjectItemResponseDtoRequester } from "./supportExternalProjectItemResponseDtoRequester";
import type { SupportExternalProjectItemResponseDtoTeam } from "./supportExternalProjectItemResponseDtoTeam";

export interface SupportExternalProjectItemResponseDto {
  allowedActions: SupportExternalProjectItemResponseDtoAllowedActionsItem[];
  /** @nullable */
  assignee: SupportExternalProjectItemResponseDtoAssignee;
  connectionId: string;
  freshness: SupportExternalProjectItemResponseDtoFreshness;
  itemId: string;
  /** @nullable */
  lastRefreshedAt: string | null;
  /** @nullable */
  latestMessageAt: string | null;
  /** @nullable */
  link?: SupportExternalProjectItemResponseDtoLink;
  linked?: boolean;
  /**
   * @maxLength 100
   * @nullable
   */
  priority: string | null;
  provider: SupportExternalProjectItemResponseDtoProvider;
  /** @maxLength 255 */
  remoteItemId: string;
  /**
   * @maxLength 255
   * @nullable
   */
  remoteKey: string | null;
  /** @nullable */
  remoteUpdatedAt: string | null;
  /** @nullable */
  remoteUrl: string | null;
  /** @nullable */
  requester: SupportExternalProjectItemResponseDtoRequester;
  /**
   * @maxLength 100
   * @nullable
   */
  status: string | null;
  /**
   * @maxLength 2000
   * @nullable
   */
  summary: string | null;
  /** @maxItems 100 */
  tags: string[];
  /** @nullable */
  team: SupportExternalProjectItemResponseDtoTeam;
  /** @minimum 1 */
  version: number;
}